+++ src/init.cpp
@@ -48,6 +48,7 @@
 #include "spork.h"
 #include "sporkdb.h"
 #include "evo/evodb.h"
+#include "tiertwo/init.h"
 #include "txdb.h"
 #include "torcontrol.h"
 #include "guiinterface.h"
@@ -947,19 +948,6 @@ static std::string ResolveErrMsg(const char * const optname, const std::string& 
     return strprintf(_("Cannot resolve -%s address: '%s'"), optname, strBind);
 }
 
-// Sets the last CACHED_BLOCK_HASHES hashes into masternode manager cache
-static void LoadBlockHashesCache(CMasternodeMan& man)
-{
-    LOCK(cs_main);
-    const CBlockIndex* pindex = chainActive.Tip();
-    unsigned int inserted = 0;
-    while (pindex && inserted < CACHED_BLOCK_HASHES) {
-        man.CacheBlockHash(pindex);
-        pindex = pindex->pprev;
-        ++inserted;
-    }
-}
-
 void InitLogging()
 {
     g_logger->m_print_to_file = !gArgs.IsArgNegated("-debuglogfile");
@@ -1742,49 +1730,8 @@ bool AppInitMain()
 
     // ********************************************************* Step 10: setup layer 2 data
 
-    uiInterface.InitMessage(_("Loading masternode cache..."));
-
-    mnodeman.SetBestHeight(chain_active_height);
-    LoadBlockHashesCache(mnodeman);
-    CMasternodeDB mndb;
-    CMasternodeDB::ReadResult readResult = mndb.Read(mnodeman);
-    if (readResult == CMasternodeDB::FileError)
-        LogPrintf("Missing masternode cache file - mncache.dat, will try to recreate\n");
-    else if (readResult != CMasternodeDB::Ok) {
-        LogPrintf("Error reading mncache.dat - cached data discarded\n");
-    }
-
-    uiInterface.InitMessage(_("Loading budget cache..."));
-
-    CBudgetDB budgetdb;
-    const bool fDryRun = (chain_active_height <= 0);
-    if (!fDryRun) g_budgetman.SetBestHeight(chain_active_height);
-    CBudgetDB::ReadResult readResult2 = budgetdb.Read(g_budgetman, fDryRun);
-
-    if (readResult2 == CBudgetDB::FileError)
-        LogPrintf("Missing budget cache - budget.dat, will try to recreate\n");
-    else if (readResult2 != CBudgetDB::Ok) {
-        LogPrintf("Error reading budget.dat - cached data discarded\n");
-    }
-
-    //flag our cached items so we send them to our peers
-    g_budgetman.ResetSync();
-    g_budgetman.ReloadMapSeen();
-
-    RegisterValidationInterface(&g_budgetman);
-
-    uiInterface.InitMessage(_("Loading masternode payment cache..."));
-
-    CMasternodePaymentDB mnpayments;
-    CMasternodePaymentDB::ReadResult readResult3 = mnpayments.Read(masternodePayments);
-
-    RegisterValidationInterface(&masternodePayments);
-
-    if (readResult3 == CMasternodePaymentDB::FileError)
-        LogPrintf("Missing masternode payment cache - mnpayments.dat, will try to recreate\n");
-    else if (readResult3 != CMasternodePaymentDB::Ok) {
-        LogPrintf("Error reading mnpayments.dat - cached data discarded\n");
-    }
+    LoadTierTwo(chain_active_height);
+    RegisterTierTwoValidationInterface();
 
     fMasterNode = gArgs.GetBoolArg("-masternode", DEFAULT_MASTERNODE);
